refactor: Extract `useIsMobile` hook and replace duplicate logic
- Created a `useIsMobile` hook to manage mobile detection logic.
- Replaced duplicated `isMobile` state and `useEffect` handling in `ExpandableDiv` and `Header` components with `useIsMobile`.
